Trying Lombok
I decided not to use lombok because I'm using Intellij Educational in which lombok is not available. Although it shortens the code with the annotations such as Getters and Setters and Constructors etc. + it helps avoiding repetitive code, in Intellij it is not a time-consuming matter anymore, thanks to the generators. @Delegate and other features unfortunately are still experimental and not recommended to use.